metrics/recovar/recovar_install_requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
appdirs==1.4.4
argcomplete==3.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==5.3.4
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
biopython==1.79
bleach==6.1.0
bokeh==3.2.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
colorcet==3.0.1
colorlog==6.7.0
colorlover==0.3.0
comm==0.1.4
contourpy==1.1.1
cryodrgn==2.3.0
cssutils==2.9.0
cufflinks==0.17.3
cupy-cuda12x==13.0.0
cycler==0.12.1
dask==2023.10.0
dataframe-image==0.2.2
datashader==0.15.2
datashape==0.5.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
equinox==0.11.3
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
fastrlock==0.8.2
filelock==3.12.4
finufft==2.1.0
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
healpy==1.16.6
holoviews==1.17.1
html2image==2.0.4.3
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==7.8.1
isoduration==20.11.0
# jax==0.4.23
# jaxlib==0.4.23+cuda12.cudnn89
jaxopt==0.8.1
jaxtyping==0.2.28
jaxwt==0.1.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.7
jupyterlab_server==2.25.0
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lineax==0.0.4
linkify-it-py==2.0.2
llvmlite==0.41.0
locket==1.0.0
lxml==4.9.3
Mako==1.3.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
matplotlib-scalebar==0.8.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
mrcfile==1.4.3
multidict==6.0.4
multipledispatch==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
nox==2023.4.22
numba==0.58.0
numpy==1.23.5
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.0
panel==1.2.3
param==1.13.0
parso==0.8.3
partd==1.4.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
pluggy==1.3.0
pooch==1.8.0
ProDy==2.4.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
pycuda==2024.1
pyerfa==2.0.1
Pygments==2.16.1
pynndescent==0.5.10
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytools==2023.1.1
pytz==2023.3.post1
pyviz_comms==3.0.0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
scikit-fmm==2023.4.2
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starfile==0.5.4
sympy==1.12
tbb==2021.10.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow-cpu==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
typeguard==2.13.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.4
uc-micro-py==1.0.2
umap-learn==0.5.4
uri-template==1.3.0
urllib3==2.0.6
virtualenv==20.24.6
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.0
widgetsnbextension==3.6.6
wrapt==1.14.1
xarray==2023.9.0
xyzservices==2023.10.0
yarl==1.9.2
zipp==3.17.0
